+++ plat/marvell/a8k/a80x0/board/dram_port.c
@@ -6,8 +6,8 @@
  */
 
 #include <arch_helpers.h>
-#include <a8k_i2c.h>
 #include <debug.h>
+#include <mentor/mi2cv.h>
 #include <mmio.h>
 #include <mv_ddr_if.h>
 #include <mvebu_def.h>
